feat: 다이어리 연동 AI 여행계획 생성, 삭제 복구(Undo), 추천 장소 개수 UI 조절 및 로딩 추가

+++ app/src/main/java/com/example/harudiary/activity/FriendRecordActivity.java
@@ -406,7 +406,7 @@ public class FriendRecordActivity extends AppCompatActivity {
         });
 
         btnComment.setOnClickListener(v -> {
-            CommentBottomSheetFragment sheet = CommentBottomSheetFragment.newInstance(activityId.intValue());
+            CommentBottomSheetFragment sheet = CommentBottomSheetFragment.newInstance(activityId);
             sheet.setOnCommentChangedListener(() -> render()); // Re-render to refresh counts
             try {
                 sheet.show(getSupportFragmentManager(), "comments_" + activityId);
